+++ Changelog.md
@@ -1,3 +1,24 @@
+## 13.519.5
+* Adds dis/advantage flags for damage rolls. From now on it is suggested that you are using:
+  * `flags.automated-confitions-5e.damage.advantage` etc
+* Properly parse formulas starting with `-`
+* Remove quotes from bonuses in en.json lang file
+* Make `itemData.type.value` available as shorthand
+* Updated Italian translation by [GregoryWarn](<https://github.com/GregoryWarn>) 🤗
+
+## 13.519.4.1
+* Backwards compatibility, fixing issues with remnant quoted parts in formulas.
+
+## 13.519.4
+* Reworked `_ac5eSafeEval`
+* Introduced new parser
+ * full ternary operators support while evaluating formulas
+ * any bonuses are folded in a final value before added to the Rolls
+* `sandbox` lazy load
+* Moved repetitive code to Functions in `ac5e-setpieces`
+* Reworked `bonus` and `set`
+* Updated Italian translation by [GregoryWarn](<https://github.com/GregoryWarn>) 🤗
+
 ## 13.519.3.1
 * Hotfix
 * Added `ac5e.info.version`